Detail Cleaning and Code Cleanup

+++ app/src/main/java/com/devchrisap/apptourism/Entities/Destiny.kt
@@ -2,7 +2,7 @@ package com.devchrisap.apptourism.Entities
 
 class Destiny(
     var _id: String,
-    var score: Double,
+    var score: Float,
     var image: String,
     var description: String,
     var cityId: Int,
